Update System Logger.
